Split angle server helpers
