+++ frontend/src/server/tools/angle.ts
@@ -1,14 +1,11 @@
 import { randomUUID } from 'crypto';
 import { ApiError, ValidationError } from '@fal-ai/client';
 import { getAngleToolEngine } from '@/config/tools-angle-engines';
-import { isDatabaseConfigured, query, type QueryExecutor, withDbTransaction } from '@/lib/db';
+import { query } from '@/lib/db';
 import { getFalClient } from '@/lib/fal-client';
-import { getResultProviderMode } from '@/lib/result-provider';
-import { ensureBillingSchema } from '@/lib/schema';
 import { computeBillingProductSnapshot } from '@/lib/billing-products';
 import { getPlatformFeeCents } from '@/lib/pricing';
-import { getUserPreferredCurrency, type Currency } from '@/lib/currency';
-import { reserveWalletChargeInExecutor } from '@/lib/wallet';
+import { getUserPreferredCurrency } from '@/lib/currency';
 import { receiptsPriceOnlyEnabled } from '@/lib/env';
 import { createImageThumbnailBatch } from '@/server/image-thumbnails';
 import { buildStoredImageRenderEntries, resolveHeroThumbFromRenders } from '@/lib/image-renders';
@@ -23,7 +20,6 @@ import type {
   AngleToolOutput,
   AngleToolRequest,
   AngleToolResponse,
-  AngleToolEngineId,
 } from '@/types/tools-angle';
 import type { PricingSnapshot } from '@/types/engines';
 import {
@@ -39,270 +35,24 @@ import {
   toAngleValidationMessage,
 } from './angle-request-utils';
 import { persistAngleOutputs } from './angle-output-persistence';
+import { AngleToolError } from './angle-error';
+import {
+  createAtomicInitialAngleJob,
+  PLACEHOLDER_THUMB,
+} from './angle-initial-job';
+import { insertAngleToolEvent, TOOL_EVENT_NAME } from './angle-event-log';
+import { recordAngleRefundReceipt, type PendingAngleReceipt } from './angle-receipts';
 
-const TOOL_EVENT_NAME = 'tool_angle_generate';
-const PLACEHOLDER_THUMB = '/assets/frames/thumb-1x1.svg';
-
-function usdToCredits(value: number | null | undefined): number | null {
-  if (typeof value !== 'number' || !Number.isFinite(value) || value <= 0) return null;
-  return Math.max(1, Math.round(value * 100));
-}
-
-export class AngleToolError extends Error {
-  status: number;
-  code: string;
-  detail?: unknown;
-
-  constructor(message: string, options?: { status?: number; code?: string; detail?: unknown }) {
-    super(message);
-    this.name = 'AngleToolError';
-    this.status = options?.status ?? 500;
-    this.code = options?.code ?? 'angle_tool_error';
-    this.detail = options?.detail;
-  }
-}
+export { AngleToolError } from './angle-error';
+export { createAngleInitialJobInExecutor } from './angle-initial-job';
 
 type RunAngleToolInput = AngleToolRequest & {
   userId: string;
 };
 
-type PendingAngleReceipt = {
-  userId: string;
-  amountCents: number;
-  currency: string;
-  description: string;
-  jobId: string;
-  surface: typeof ANGLE_SURFACE;
-  billingProductKey: string;
-  snapshot: PricingSnapshot;
-  applicationFeeCents: number | null;
-  vendorAccountId: string | null;
-};
-
-type ProvisionalAngleJobInsert = {
-  jobId: string;
-  userId: string;
-  billingProductKey: string;
-  engineId: AngleToolEngineId;
-  engineLabel: string;
-  durationSec: number;
-  promptSummary: string;
-  finalPriceCents: number;
-  pricingSnapshotJson: string;
-  settingsSnapshotJson: string;
-  currency: string;
-  vendorAccountId: string | null;
-};
-
-type CreateAngleInitialJobParams = {
-  userId: string;
-  jobId: string;
-  description: string;
-  amountCents: number;
-  currency: string;
-  billingProductKey: string;
-  pricingSnapshotJson: string;
-  applicationFeeCents: number | null;
-  vendorAccountId: string | null;
-  engineId: AngleToolEngineId;
-  engineLabel: string;
-  requestedOutputCount: number;
-  promptSummary: string;
-  settingsSnapshotJson: string;
-  preferredCurrency: Currency | null;
-};
-
-async function recordAngleRefundReceipt(receipt: PendingAngleReceipt, label: string, priceOnly: boolean) {
-  try {
-    await query(
-      `INSERT INTO app_receipts (
-         user_id,
-         type,
-         amount_cents,
-         currency,
-         description,
-         job_id,
-         surface,
-         billing_product_key,
-         pricing_snapshot,
-         application_fee_cents,
-         vendor_account_id,
-         platform_revenue_cents,
-         destination_acct
-       )
-       VALUES ($1,'refund',$2,$3,$4,$5,$6,$7,$8::jsonb,$9,$10,$11,$12)
-       ON CONFLICT DO NOTHING`,
-      [
-        receipt.userId,
-        receipt.amountCents,
-        receipt.currency,
-        label,
-        receipt.jobId,
-        receipt.surface,
-        receipt.billingProductKey,
-        JSON.stringify(receipt.snapshot),
-        priceOnly ? null : 0,
-        priceOnly ? null : receipt.vendorAccountId,
-        priceOnly ? null : 0,
-        priceOnly ? null : receipt.vendorAccountId,
-      ]
-    );
-  } catch (error) {
-    console.warn('[tools/angle] failed to record refund receipt', error);
-  }
-}
-
-async function insertProvisionalAngleJob(executor: QueryExecutor, params: ProvisionalAngleJobInsert) {
-  await executor.query(
-    `INSERT INTO app_jobs (
-       job_id,
-       user_id,
-       surface,
-       billing_product_key,
-       engine_id,
-       engine_label,
-       duration_sec,
-       prompt,
-       thumb_url,
-       preview_frame,
-       status,
-       progress,
-       final_price_cents,
-       pricing_snapshot,
-       settings_snapshot,
-       currency,
-       vendor_account_id,
-       payment_status,
-       visibility,
-       indexable,
-       provisional
-     )
-     VALUES (
-       $1,$2,$3,$4,$5,$6,$7,$8,$9,$10,'pending',0,$11,$12::jsonb,$13::jsonb,$14,$15,'paid_wallet','private',FALSE,TRUE
-     )`,
-    [
-      params.jobId,
-      params.userId,
-      ANGLE_SURFACE,
-      params.billingProductKey,
-      params.engineId,
-      params.engineLabel,
-      params.durationSec,
-      params.promptSummary,
-      PLACEHOLDER_THUMB,
-      PLACEHOLDER_THUMB,
-      params.finalPriceCents,
-      params.pricingSnapshotJson,
-      params.settingsSnapshotJson,
-      params.currency,
-      params.vendorAccountId,
-    ]
-  );
-}
-
-export async function createAngleInitialJobInExecutor(
-  executor: QueryExecutor,
-  params: CreateAngleInitialJobParams
-): Promise<void> {
-  const reserveResult = await reserveWalletChargeInExecutor(
-    executor,
-    {
-      userId: params.userId,
-      amountCents: params.amountCents,
-      currency: params.currency,
-      description: params.description,
-      jobId: params.jobId,
-      surface: ANGLE_SURFACE,
-      billingProductKey: params.billingProductKey,
-      pricingSnapshotJson: params.pricingSnapshotJson,
-      applicationFeeCents: params.applicationFeeCents,
-      vendorAccountId: params.vendorAccountId,
-      stripePaymentIntentId: null,
-      stripeChargeId: null,
-    },
-    { preferredCurrency: params.preferredCurrency }
-  );
-
-  if (!reserveResult.ok) {
-    if (reserveResult.errorCode === 'currency_mismatch') {
-      const lockedCurrency = (reserveResult.preferredCurrency ?? 'usd').toUpperCase();
-      throw new AngleToolError(`Wallet currency locked to ${lockedCurrency}. Contact support to request a change.`, {
-        status: 409,
-        code: 'wallet_currency_mismatch',
-        detail: { lockedCurrency },
-      });
-    }
-
-    throw new AngleToolError('Insufficient wallet balance for this angle run.', {
-      status: 402,
-      code: 'insufficient_wallet_funds',
-      detail: {
-        balanceCents: reserveResult.balanceCents,
-        requiredCents: Math.max(0, params.amountCents - reserveResult.balanceCents),
-      },
-    });
-  }
-
-  await insertProvisionalAngleJob(executor, {
-    jobId: params.jobId,
-    userId: params.userId,
-    billingProductKey: params.billingProductKey,
-    engineId: params.engineId,
-    engineLabel: params.engineLabel,
-    durationSec: params.requestedOutputCount,
-    promptSummary: params.promptSummary,
-    finalPriceCents: params.amountCents,
-    pricingSnapshotJson: params.pricingSnapshotJson,
-    settingsSnapshotJson: params.settingsSnapshotJson,
-    currency: params.currency,
-    vendorAccountId: params.vendorAccountId,
-  });
-}
-
-async function createAtomicInitialAngleJob(params: CreateAngleInitialJobParams): Promise<void> {
-  try {
-    await withDbTransaction(async (executor) => {
-      await executor.query(`SELECT pg_advisory_xact_lock(hashtext($1))`, [params.jobId]);
-      await createAngleInitialJobInExecutor(executor, params);
-    });
-  } catch (error) {
-    if (error instanceof AngleToolError) {
-      throw error;
-    }
-
-    throw new AngleToolError('Failed to save angle job.', {
-      status: 500,
-      code: 'job_persist_failed',
-      detail: error instanceof Error ? error.message : error,
-    });
-  }
-}
-
-async function insertToolEvent(params: {
-  jobId: string;
-  engineId: AngleToolEngineId;
-  providerJobId?: string | null;
-  payload: Record<string, unknown>;
-}) {
-  if (!isDatabaseConfigured()) return;
-
-  try {
-    await ensureBillingSchema();
-    await query(
-      `INSERT INTO fal_queue_log (job_id, provider, provider_job_id, engine_id, status, payload)
-       VALUES ($1,$2,$3,$4,$5,$6::jsonb)`,
-      [
-        params.jobId,
-        getResultProviderMode(),
-        params.providerJobId ?? null,
-        params.engineId,
-        TOOL_EVENT_NAME,
-        JSON.stringify(params.payload),
-      ]
-    );
-  } catch (error) {
-    console.warn('[tools/angle] failed to persist event log', error);
-  }
+function usdToCredits(value: number | null | undefined): number | null {
+  if (typeof value !== 'number' || !Number.isFinite(value) || value <= 0) return null;
+  return Math.max(1, Math.round(value * 100));
 }
 
 export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolResponse> {
@@ -523,7 +273,7 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
       )
     );
 
-    await insertToolEvent({
+    await insertAngleToolEvent({
       jobId,
       engineId: engine.id,
       providerJobId: requestId,
@@ -628,7 +378,7 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
       console.warn('[tools/angle] failed to mark job as failed', updateError);
     }
 
-    await insertToolEvent({
+    await insertAngleToolEvent({
       jobId,
       engineId: engine.id,
       providerJobId,
