+++ tests/angle-server-architecture.test.ts
@@ -7,10 +7,18 @@ const root = process.cwd();
 const serverPath = join(root, 'frontend/src/server/tools/angle.ts');
 const requestUtilsPath = join(root, 'frontend/src/server/tools/angle-request-utils.ts');
 const outputPersistencePath = join(root, 'frontend/src/server/tools/angle-output-persistence.ts');
+const angleErrorPath = join(root, 'frontend/src/server/tools/angle-error.ts');
+const angleReceiptsPath = join(root, 'frontend/src/server/tools/angle-receipts.ts');
+const angleInitialJobPath = join(root, 'frontend/src/server/tools/angle-initial-job.ts');
+const angleEventLogPath = join(root, 'frontend/src/server/tools/angle-event-log.ts');
 
 const serverSource = readFileSync(serverPath, 'utf8');
 const requestUtilsSource = readFileSync(requestUtilsPath, 'utf8');
 const outputPersistenceSource = readFileSync(outputPersistencePath, 'utf8');
+const angleErrorSource = readFileSync(angleErrorPath, 'utf8');
+const angleReceiptsSource = readFileSync(angleReceiptsPath, 'utf8');
+const angleInitialJobSource = readFileSync(angleInitialJobPath, 'utf8');
+const angleEventLogSource = readFileSync(angleEventLogPath, 'utf8');
 
 test('angle server delegates request and provider normalization helpers', () => {
   assert.ok(existsSync(requestUtilsPath), 'angle request helpers should live in a server-local utility module');
@@ -42,7 +50,7 @@ test('angle server delegates request and provider normalization helpers', () => 
   assert.doesNotMatch(serverSource, /normalizeMediaUrl/, 'Fal output URL normalization belongs in angle-request-utils.ts');
 
   const lineCount = serverSource.split('\n').length;
-  assert.ok(lineCount <= 700, `angle server orchestrator should stay below 700 lines after helper extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 460, `angle server orchestrator should stay below 460 lines after helper extraction, got ${lineCount}`);
 });
 
 test('angle server delegates output persistence to a focused helper', () => {
@@ -96,4 +104,68 @@ test('angle request utils expose the expected pure helper contract', () => {
 
   assert.match(requestUtilsSource, /function normalizeFalUrl\(/, 'Fal URL normalization should be private to request helpers');
   assert.match(requestUtilsSource, /function toAngleOutput\(/, 'provider output mapping should be private to request helpers');
+});
+
+test('angle server delegates wallet, receipt, and event-log responsibilities', () => {
+  for (const [path, label] of [
+    [angleErrorPath, 'angle error helper'],
+    [angleReceiptsPath, 'angle receipt helper'],
+    [angleInitialJobPath, 'angle initial job helper'],
+    [angleEventLogPath, 'angle event log helper'],
+  ] as const) {
+    assert.ok(existsSync(path), `${label} should live in a server-local helper module`);
+  }
+
+  for (const importName of ['angle-error', 'angle-receipts', 'angle-initial-job', 'angle-event-log']) {
+    assert.match(serverSource, new RegExp(`from '\\./${importName}'`), `angle server should import ${importName}`);
+  }
+
+  assert.match(serverSource, /export \{ AngleToolError \} from '\.\/angle-error'/, 'legacy error export should stay stable');
+  assert.match(
+    serverSource,
+    /export \{ createAngleInitialJobInExecutor \} from '\.\/angle-initial-job'/,
+    'legacy initial-job export should stay stable'
+  );
+
+  for (const implementationPattern of [
+    /export class AngleToolError/,
+    /reserveWalletChargeInExecutor/,
+    /withDbTransaction/,
+    /async function insertProvisionalAngleJob/,
+    /async function recordAngleRefundReceipt/,
+    /async function insertToolEvent/,
+  ]) {
+    assert.doesNotMatch(serverSource, implementationPattern, 'angle.ts should stay focused on run orchestration');
+  }
+
+  assert.match(angleErrorSource, /export class AngleToolError/, 'angle-error should own the public error class');
+  assert.match(angleReceiptsSource, /export type PendingAngleReceipt/, 'angle-receipts should export pending receipt data');
+  assert.match(
+    angleReceiptsSource,
+    /export async function recordAngleRefundReceipt/,
+    'angle-receipts should own refund receipt persistence'
+  );
+  assert.match(angleInitialJobSource, /export const PLACEHOLDER_THUMB/, 'angle-initial-job should own placeholder metadata');
+  assert.match(
+    angleInitialJobSource,
+    /export async function createAngleInitialJobInExecutor/,
+    'angle-initial-job should export the testable transaction inner contract'
+  );
+  assert.match(
+    angleInitialJobSource,
+    /export async function createAtomicInitialAngleJob/,
+    'angle-initial-job should export atomic initial-job persistence'
+  );
+  assert.match(
+    angleInitialJobSource,
+    /reserveWalletChargeInExecutor/,
+    'angle-initial-job should own wallet reservation preflight'
+  );
+  assert.match(angleInitialJobSource, /withDbTransaction/, 'angle-initial-job should own the transaction boundary');
+  assert.match(angleEventLogSource, /export const TOOL_EVENT_NAME/, 'angle-event-log should own the event name');
+  assert.match(
+    angleEventLogSource,
+    /export async function insertAngleToolEvent/,
+    'angle-event-log should own Fal queue event persistence'
+  );
 });
